# api/openapi.yaml
openapi: "3.0.3"
info:
  title: Promptsheon API
  description: Version Control System for AI agent intelligence
  version: "1.0.0"
  contact:
    name: Promptsheon
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "API key prefixed with ps_. Send as: Authorization: Bearer ps_..."
  schemas:
...
    Agent:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        steps:
          type: array
          items:
            $ref: "#/components/schemas/AgentStep"
        tools:
          type: array
          items:
            $ref: "#/components/schemas/ToolRef"
        status:
          type: string
        cas_hash:
          type: string
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AgentStep:
      type: object
      properties:
        id:
          type: string
        prompt_id:
          type: string
        prompt_hash:
          type: string
        depends_on:
          type: array
          items:
            type: string
        tool_calls:
          type: array
          items:
            $ref: "#/components/schemas/ToolCall"
        output_key:
          type: string
    ToolRef:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [http, shell, json_transform, prompt_call]
        config:
          type: object
    ToolCall:
      type: object
      properties:
        tool:
          type: string
        input:
          type: object
        output:
          type: object
        error:
          type: string
        latency_ms:
          type: integer
...
    APIKey:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        name:
          type: string
        key_prefix:
          type: string
        role:
          type: string
        expires_at:
          type: string
          format: date-time
        last_used:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        revoked:
          type: boolean
paths:
...
  /api/v1/apikeys:
    post:
      summary: Create API key
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, name, role]
              properties:
                user_id:
                  type: string
                name:
                  type: string
                role:
                  type: string
                  enum: [admin, writer, reader]
                expires_at:
                  type: string
                  format: date-time
      responses:
        "201":
          description: Key created (key shown once)
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:
                    type: string
                  record:
                    $ref: "#/components/schemas/APIKey"
    get:
      summary: List API keys for a user
      tags: [Auth]
      security: []
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Key list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/APIKey"
...
  /api/v1/agents:
    get:
      summary: List agents
      tags: [Agents]
      responses:
        "200":
          description: Agent list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
    post:
      summary: Create agent
      tags: [Agents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
                steps:
                  type: array
                  items:
                    $ref: "#/components/schemas/AgentStep"
                tools:
                  type: array
                  items:
                    $ref: "#/components/schemas/ToolRef"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /api/v1/agents/{id}:
    get:
      summary: Get agent
      tags: [Agents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
    put:
      summary: Update agent
      tags: [Agents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                steps:
                  type: array
                  items:
                    $ref: "#/components/schemas/AgentStep"
                tools:
                  type: array
                  items:
                    $ref: "#/components/schemas/ToolRef"
                status:
                  type: string
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
    delete:
      summary: Delete agent
      tags: [Agents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /api/v1/agents/{id}/versions:
    get:
      summary: List agent versions
      tags: [Agents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Agent versions
          content:
            application/json:
              schema:
                type: object
...